requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0

# Content Management
django-ckeditor==6.7.0
django-mptt==0.14.0
django-taggit==4.0.0

# File Handling
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7

# Task Queue
celery==5.3.4
redis==5.0.1

# Environment Management
python-decouple==3.8

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-slugify==8.0.1
reportlab==4.0.7